backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
elevenlabs
faster-whisper
torch
numpy
google-genai
pydub
ollama
